automation/requirements.txt
telethon>=1.36.0
tqdm>=4.66.0
pyyaml>=6.0.1
